# openapi.yaml
openapi: 3.0.0
info:
  title: StudentHub API
  description: Platform for Egyptian students – job matching, video CVs, AI search, discount vouchers
  version: 1.0.0
servers:
  - url: https://tolabco.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /profile:
    get:
      security: [{ bearerAuth: [] }]
      summary: Get authenticated user profile
      responses:
        '200': { description: Profile data }
        '401': { description: Unauthorized }
    put:
      security: [{ bearerAuth: [] }]
      summary: Update student profile (student only)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name: { type: string }
                date_of_birth: { type: string, format: date }
                city: { type: string }
                profession_category: { type: string }
                skills: { type: array, items: { type: string } }
      responses:
        '200': { description: Profile updated }
        '403': { description: Only students can update }
  /jobs:
    get:
      security: [{ bearerAuth: [] }]
      summary: List jobs with search, filters, pagination
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: job_type, in: query, schema: { type: string } }
        - { name: location_city, in: query, schema: { type: string } }
        - { name: is_remote, in: query, schema: { type: boolean } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
      responses:
        '200': { description: Jobs with pagination }
    post:
      security: [{ bearerAuth: [] }]
      summary: Create a job (employer only)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                description: { type: string }
                job_type: { type: string }
                location_city: { type: string }
                is_remote: { type: boolean }
                salary_range: { type: string }
      responses:
        '201': { description: Job created }
        '403': { description: Only employers }
    put:
      security: [{ bearerAuth: [] }]
      summary: Edit a job (employer only)
      parameters:
        - { name: jobId, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                job_type: { type: string }
                location_city: { type: string }
                is_remote: { type: boolean }
                salary_range: { type: string }
      responses:
        '200': { description: Job updated }
    delete:
      security: [{ bearerAuth: [] }]
      summary: Delete a job (employer only)
      parameters:
        - { name: jobId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Job deleted }
...
  /get-upload-url:
    get:
      security: [{ bearerAuth: [] }]
      summary: Get pre-signed URL for video upload (student only)
      parameters:
        - { name: extension, in: query, schema: { type: string, default: mp4 } }
      responses:
        '200': { description: Returns uploadUrl and key }
  /set-video-cv:
    post:
      security: [{ bearerAuth: [] }]
      summary: Save video reference after upload (student only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key]
              properties:
                key: { type: string }
      responses:
        '200': { description: Video saved }
  /cv/{shortCode}:
    get:
      summary: Public CV page (no auth)
      parameters:
        - { name: shortCode, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Student profile with video URL }
        '404': { description: Not found }
...
  /vouchers:
    get:
      security: [{ bearerAuth: [] }]
      summary: List active vouchers
      responses:
        '200': { description: Vouchers list }
    post:
      security: [{ bearerAuth: [] }]
      summary: Create a voucher (outlet only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, discount_percent]
              properties:
                title: { type: string }
                description: { type: string }
                discount_percent: { type: integer }
                start_date: { type: string, format: date }
                end_date: { type: string, format: date }
                max_redemptions: { type: integer }
      responses:
        '201': { description: Voucher created }
  /redeem/{voucherId}:
    post:
      security: [{ bearerAuth: [] }]
      summary: Redeem a voucher (student only)
      parameters:
        - { name: voucherId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Redeemed successfully }
...